COODE REFACTORING: code refactoring and renaming
