+++ src/components/authorization/EditAPIToken.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function EditAPIToken() {
+    return <div>EditAPIToken</div>
+}
+
+export default EditAPIToken
