+++ src/components/authorization/GenerateActionButton.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function GenerateActionButton({ handleGenerateRowActionButton }) {
+    return (
+        <button onClick={() => handleGenerateRowActionButton('create')} className="add-link cta flex cursor">
+            Generate new token
+        </button>
+    )
+}
+
+export default GenerateActionButton
