+++ src/components/globalConfigurations/GlobalConfiguration.tsx
@@ -28,7 +28,7 @@ const Project = lazy(() => import('../project/ProjectList'))
 const UserGroup = lazy(() => import('../userGroups/UserGroup'))
 const SSOLogin = lazy(() => import('../login/SSOLogin'))
 const CustomChartList = lazy(() => import('../CustomChart/CustomChartList'))
-const ApiTokens = lazy(() => import('../authorization/ApiTokens'))
+const ApiTokens = lazy(() => import('../authorization/ApiTokens.component'))
 
 export default function GlobalConfiguration(props) {
     const location = useLocation()
@@ -163,7 +163,7 @@ function NavItem({ hostURLConfig, serverMode }) {
         { name: 'SSO login services', href: URLS.GLOBAL_CONFIG_LOGIN, component: SSOLogin, isAvailableInEA: true },
         {
             name: 'API token',
-            href: `${URLS.GLOBAL_CONFIG_AUTH}${URLS.GLOBAL_CONFIG_API}`,
+            href: `${URLS.GLOBAL_CONFIG_AUTH}/list`,
             component: ApiTokens,
             isAvailableInEA: true,
         },
